Adds new video component.
This was something we had custom built as a markdown / React component for the calcite.esri.com site, and figured it'd be a good addition here.

Has a set of configurable options for:
- allowing fullscreen
- allowing scrubbing
- playing on hover (and hiding controls / progress) - this is the use case from the calcite.esri.com documentation

It  displays the calcite loader while the video loads and has events for play / pause.

Still need to add tests and run through accessibility, and allow users to pass strings for play / pause / restart / volume / fullscreen button labels.