fix(alert): ensure `border-radius` is consistent for prescribed `slots` (#6368)
**Related Issue:** #6348

## Summary
The bottom left corner radius was missing when the icon was present.
Upon inspection, it was clear that slots added around content don't have
the proper radius set and end up covering the radius on the host. For
each occasion the slot is placed against either end of the host, I
adjusted the relevant radius to match the host container.

Because the close button is now always present, the cases addressed are
the right side of the close button and the left side of the content (no
icon) along with the left side of the icon (when placed in front of the
content).

Added a screenshot for the radius with options where `icon` is present
and absent.