fix(tree-item): preserves consistent height with or w/t actions-end  (#6403)
**Related Issue:** #6361, #3127


## Summary
Moved the `actions-end` slot outside of the `node-container` layout and
placed it next to it instead, similar to `notice`. This way the padding
on the `node-container` does not add to the action-ends.

Expected consistent height on the `tree-item` when bringing in
`actions-end` can be preserved like this (here to show designers how
this can be demo-ed to end-users).

s (24) tree-item: s slotted component + s action
m (32) tree-item: m slotted component + m action
l (48) tree-item: l slotted component + l action


https://user-images.githubusercontent.com/19231036/216716464-8413540a-85f5-4183-a2b1-b67652ae1384.mov


We need to follow up on clipping the action:

- We’re not normally clipping actions to accommodate component heights,
it’s inconsistent with how we normally handle slots. So here I’m
clipping the actions for the time being and marking this internally as a
known issue (Figma and implementation side)
- Large clips outline: clips top and bottom, will be added to the note
above.