feat(modal): provides `content-top` and `content-bottom` slots (#6490)
**Related Issue:** #4800

## Summary
Provides `content-top` and `content-bottom` slots. 

For simplicity, we matched the `padding` on the `content-top` and
`content-bottom` to the `content`, with one public prop for the latter.
This can further be developed if requested.

We decided to move forward with `content-top` and `content-bottom`
naming, as this should provide enough context and also aligns with
existing `<x>-start/<x>-end` naming.