fix(alert, combobox, dropdown, input-date-picker, popover, tooltip): prefers-reduced-motion no longer prevents open/close components from emitting before + open/close events (#6605)
**Related Issue:** #6582 

## Summary

This restores a workaround to get open/close components to emit
before+open/close events when `prefers-reduced-motion` is enabled. Also,
a Sass utility was added help scale durations for animations that
require fine-tuning. This technically restores some animation, but it is
very reduced:

### `calcite-block`

![block-reduced-motion](https://user-images.githubusercontent.com/197440/225195863-f64faee5-ba13-4f6c-925a-2de1caef0718.gif)

### `calcite-loader`

![loader-reduced-motion](https://user-images.githubusercontent.com/197440/225195875-a938f628-1a19-48a8-b6a8-f6967bc05a21.gif)

### `calcite-progress`

![progress-reduced-motion](https://user-images.githubusercontent.com/197440/225195853-7509681d-5c1d-46f7-b103-3dacf3c5bb6b.gif)

I _think_ this still aligns with the expected behavior of
`prefers-reduced-motion`, but we should check to just in case. cc
@geospatialem