ci: add label to skip visual snapshots (#6935)
**Related Issue:** #

## Summary
Adds a `skip visual snapshots` label which can be used for PRs that
don't need to run chromatic. The new label will skip the need to run visual snapshots. When a PR has that label, the PR check will automatically show up as successful every time you push. 

Conversely, the `pr ready for visual snapshots` label is used to run the snapshots, and it only does so during the act of adding it. So if a PR has the ready label and you push and need to rerun the snapshots, you need to toggle the label to get it to run again. This is to prevent unnecessary chromatic runs for cleanup commits and other stuff because we are charged per snapshot.

Without using either label, the PR check will be in the yellow "Expected" state, which blocks merging since it is a required check. So you need to use one label or the other, depending on if you want chromatic to run or not.