feat: tailwind to use calcite-design-tokens instead of calcite-colors (#6884)
## Summary
Calcite Design Tokens provides the same color variables as Calcite
Colors. This should not change the styling of calcite-components or the
experience for the design system consumers. This sets us up for success
moving forward as Calcite Components may now align to the style tokens
established by the Calcite design team.