chore: remove designated Element comment block and move  to Private Props (#7541)
**Related Issue:** #N/A

## Summary
Remove the designated `Element` comment block and move `@Element()` to
the `Private Props` block.

We don't have an automated way of enforcing a designated slot for just
the `Element` at the moment. `@Element()` is not really that different
from any other private prop in a component, no different from `@State()`
for example, and it's one less thing to maintain or keep track of.