fix(loader): tweak loading animations to work in Safari (#7564)
**Related Issue:** #6383 

## Summary

This revisits how the loader animates its different segments to play
nicely across all browsers. It seems like the percentages calculated
[here](https://github.com/Esri/calcite-design-system/blob/main/packages/calcite-components/src/components/loader/loader.scss#L179-L181)
are causing Safari to interrupt the animation before resetting (there
are some [existing](https://bugs.webkit.org/show_bug.cgi?id=258897)
[issues](https://bugs.webkit.org/show_bug.cgi?id=249307) on this
behavior). It comes close to the original, but it's not identical, so
I'm open to suggestions on how to improve this. cc @driskull
@macandcheese @paulcpederson

Also, the `loader-clockwise` (rotating) animation is now applied to the
SVG container to avoid it being overwritten by the `loader-offset-#`
animations.

This also fixes the determinate loader logic that mimics a loader
updating dynamically (demo page).