fix(combobox): fix filtering to avoid showing unrelated items (#7704)
**Related Issue:** #6814 

## Summary

This fixes an issue with the logic matching items to the internal
filtered data. Filtered data corresponds to items, so both `label` and
`value` need to match vs one or the other.

Also, hidden groups are visually hidden now. They were being displayed
because the host was missing the base `hidden` styles.