feat(card): Add component tokens (#8605)
**Related Issue:** #7180 

## Summary

Adds the following public component css properties:
`--calcite-card-background-color`: Specifies the background color of the
component.
`--calcite-card-border-color`: Specifies the border color of the
component.
`--calcite-card-box-shadow`: Specifies the box shadow of the component.
`--calcite-card-corner-radius`: Specifies the corner radius of the
component.

Adds a Chromatic test and a demo page example. Also replaces broken
image links in local demo.