fix(table): Prevent duplicate scrollbars in certain browsers (#8962)
**Related Issue:** #8947 

## Summary
Fixes an issue in Chrome and Safari where double scrollbars would
appear. As mentioned in the issue, it does seem to have to do with the
conditional rendering of the assistive technology text. Firefox was
unaffected. Honestly kind of a strange bug, I don't think (?) this has
come up before with the AT hidden class we use.

cc @geospatialem to verify this still works as expected in JAWS / NVDA -
the issue was introduced with the recent addition of the new
`interactionMode` property, so let's verify this doesn't have a negative
effect on AT.