ci: prevent outdated types from blocking release (#9011)
**Related Issue:** #9008

## Summary

Commit changes to `components.d.ts` after building for release to
prevent lerna/npm publishing errors due to an unclean working tree. The
commit will be discarded once the workflow run's container is destroyed.

### Background

The `components.d.ts` file is tracked as a temporary workaround for
https://github.com/ionic-team/stencil/issues/3239

To keep the types updated, `components.d.ts` is added to the index in a
pre-commit hook.


https://github.com/Esri/calcite-design-system/blob/2bb1ee24bb317c95e2bef8165c75583fe98ad714/.husky/pre-commit#L7-L13

However, the file can still be outdated if people make changes and don't
rebuild before creating another commit. This is mostly an issue with
small doc updates, like the PR linked above. If that happens, the
release will fail due to an unclean working tree.