feat(table): Add `selection-display` property (#9355)
**Related Issue:** #9353 

## Summary
This adds a `selectionDisplay` property to Table. It defaults to "top"
(the current behavior), and adds a "none" option. This will hide all
selection chips and clearance affordances. It will also hide any slotted
content in `selection-actions` slot, which is noted in the doc.

We've planned on adding further customization ability to this selection
mode UI (make it sticky, floating on bottom, etc.), and that can still
occur as planned
https://github.com/Esri/calcite-design-system/issues/8914, by adding
values to this new property.