chore: Update placeholder images in stories and demos (#9395)
**Related Issue:** #9361 , #9392 

## Summary
Removes served images from stories and local demo. We keep them in `.md`
file, as these need to be hosted somewhere in order to work if the
samples are used in an external location. cc @driskull

Ideally, we could use map tiles or something
(https://codepen.io/mac_and_cheese/pen/XWwXXqx?editors=1000) for these,
but not sure how reliable those are.