ci(provenance): add missing options from setup doc
Update the deployment GitHub Actions according to steps in the npm doc,
which I missed because I was looking at the 3rd party section below it.

The GitHub doc about provenance doesn't mention permissions, so it might
not be required.

We had originally hardcoded an older version of Ubuntu in #5942 due to
errors, but it was resolved in Stencil v4.

npm-doc: https://docs.npmjs.com/generating-provenance-statements#publishing-packages-with-provenance-via-github-actions
github-doc: https://docs.github.com/en/actions/publishing-packages/publishing-nodejs-packages#publishing-packages-to-the-npm-registry