fix: fix regression causing open/close events from emitting in proper order (#9560)
**Related Issue:** #9559

## Summary

This fixes `openCloseComponent.onToggleOpenCloseComponent` to emit
beforeOpen/Close events when the associated transition starts and emits
open/close when the transition ends.

This regression was introduced by
https://github.com/Esri/calcite-design-system/pull/9341 and would cause
`beforeOpen`/`open` or `beforeClose`/`close` to emit immediately after
another after the transition was done.

### Notable changes

* adds spec test to add coverage for `onToggleOpenCloseComponent`
* enhances `whenAnimationDone`/`whenTransitionDone` to accept callbacks
for both start and end phases
* exposes `readTask` from `openCloseComponent` to allow for stubbing
(stubbing Stencil's utility is not possible because the core module
getters are non-configurable)
* extracts animation/transition helpers for spec tests to individual
modules