ci: resolve pre-release commit but not release (#9631)
**Related Issue:** #9578

## Summary

The PR linked above introduced a bug in the pre-release automation
script, which
abstracted the steps for use on multiple branches. The bug resulted in
some
"build: update types and package-lock" commits on `dev` that versioned
`next` but didn't actually release.

I brought the state of the repo back to before those commits, and fixed
the
issues. I also removed the hotfix pre-release logic, so pre-releases can
only occur
on `dev` and `next`.