ci(chromatic): reset baseline to dev (#9746)
# Summary

Reset the Chromatic baseline of the epic branch due to invalid snapshots
being accepted. The baseline auto accept is disabled until the snapshots
issues are resolved.

**Snapshots need to be approved/rejected on both the pull request and
the squashed commit on `epic/7180-component-tokens`. Things can get out
of sync again if the new baseline on the epic branch is not
accepted/rejected after installing a PR.**

ref:
https://www.chromatic.com/docs/branching-and-baselines/#what-if-you-accidentally-reset-your-baselines-and-are-seeing-onl