feat: provide info message on stamped version instead of warning (#9739)
**Related Issue:** #9721

## Summary

This updates the message displayed when components load to show the
version, build date and revision (similar to Maps SDK). It will log an
info message instead of a warning as it could be misinterpreted and
would also show up multiple times when a bundled `calcite-components`
version would be overridden.

**Note**: `console.info` messages will be omitted by default across all
tests to avoid noise