chore: add 'start:components' npm script (#9942)
**Related Issue:** #9938

## Summary

Running `npm start` from the root of the monorepo will run `npm start`
for all of the packages that have a `start` script. Therefore, `npm
start` will launch the icon browser in addition to the component demos,
now that the `@esri/calcite-ui-components` lives in the monorepo.

This seems to cause timing issues on Windows where the icons are copied
to the components before they are all built.

To only launch the component demos, use `npm run start:components`.