feat(tooltip): allow focusing and clicking on interactive elements within a tooltip (#9914)
**Related Issue:** #6298

## Summary

- Allows placing interactive elements in tooltip elements with the
following caveats
- It should be advised against placing interactive elements within a
tooltip because there is no a11y guidelines for accessing interactive
elements within a tooltip role. Ideally a popover should be used for
interactive elements.
- No focus trapping or focus will be brought to interactive elements
within the tooltip.
- You can tab to a focusable element within a tooltip if the tooltip
with the focusable element is the next focusable element within the DOM
order. Otherwise, you could listen for the tooltip to be open and focus
on the first interactive element if you wanted to do so.
- add test 
- add story