fix: ensure `beforeOpen`/`open` and `beforeClose`/`close` events emit properly (#9958)
**Related Issue:** #9641, #9315

## Summary

This addresses some issues that prevented `beforeOpen`/`open` and
`beforeClose`/`close` events from emitting as expected.

### Notable changes:

* Removes redundant/unused transitions in favor of ones used for
open/close transitions
* Updates `input-date-picker` to set `open` via a prop instead of a
attribute to avoid the transition starting before
`onToggleOpenCloseComponent` was called. We could refactor this if
`popover` and similar components use a different, internal, attribute to
control transitions.
* Updates tests that would now have a transition when opening/closing
* Remove unnecessary `OpenCloseComponent` implementation from
`input-time-picker` as it can delegate to the internal `popover`.
* Updates `block` to use on margin transition to determine when it is
open or closed. I will create a refactor issue to refactor the
open/close transition, focusing on the content, and will update the
implementation to use this instead.
* Refactored `notice`'s styles to properly transition when
opened/closed.
* Added missing `openClose` tests to `sheet` and `tooltip` E2E test
suites.
* Added option to `openClose` test helper for tests that are expected to
use the fallback due to the normal transition start/end events not being
able to fire properly.
* Improved `openClose` test helper to consider timing between
beforeOpen/Close and open/close events.
* Fixed `whenTransitionOrAnimationDone` issue that prevented the correct
duration to be associated with the open/close transition.