ci(actions): fix adding labels action failing on issues without "Esri team" line (#10273)
**Related Issue:** #10011 

## Summary
The original issue was for fixing the action not successfully adding the
priority label to a newly created issue. It looks like the reason that
wasn't working was because the action would fail before getting to that
step.

The failure was caused by the `addEsriProductLabel` step of the action
expecting there to always be an "Esri team" line in the issue body but
that's not the case for all issue types.

This is just a small change to stop the failure and instead log a
message when the "Esri team" is not present on an issue and continue to
the next steps.