refactor: do not use Host in functional components
Lumina does not have a `<Host>` element, but there is an equivalent
syntax.

The codemod can take care of migrating to the alternative syntax, as
long as the `<Host>` element appears in the `render()` methods.

This PR moves the `<Host>` element from the `<List>` functional
component into the parent `render()` methods.

With this change, all `<Host>` elements in Calcite's codebase are
migrated automatically.