ci: compile estimate totals per milestone (#10442)
**Related Issue:** #8359

## Summary

Add an action that creates a spreadsheet and JSON file with the total
estimates per milestone. The action runs when an issue closes and can be
manually dispatched. The files are saved as artifacts in the workflow
run. I'll add a wiki page about running the workflow and retrieving the
data once it's live.

It might be because I'm a GraphQL noob, but Monday's API reference was
not very helpful. I can give the integration another try if this is not
ergonomic enough.