feat(input): Add component tokens (#10820)
**Related Issue:** #7180 

This will also serve as a pattern for related component like Input Text,
Input Number, etc.

## Summary

`--calcite-input-background-color`: Specifies the background color of
the component.
`--calcite-input-border-color`: Specifies the border color of the
component.
`--calcite-input-corner-radius`: Specifies the corner radius of the
component.
`--calcite-input-shadow`: Specifies the shadow of the component.

`--calcite-input-icon-color`: Specifies the icon color of the component.
`--calcite-input-text-color`: Specifies the text color of the component.
`--calcite-input-placeholder-text-color`: Specifies the color of
placeholder text in the component.

`--calcite-input-actions-background-color`: Specifies the background
color of Input's `clearable` and `number-button-type` elements.
`--calcite-input-actions-background-color-hover`: Specifies the
background color of Input's `clearable` and `number-button-type`
elements when hovered.
`--calcite-input-actions-background-color-press`: Specifies the
background color of Input's `clearable` and `number-button-type`
elements when pressed.
`--calcite-input-actions-icon-color`: Specifies the icon color of
Input's `clearable` and `number-button-type` elements.
`--calcite-input-actions-icon-color-hover`: Specifies the icon color of
Input's `clearable` and `number-button-type` elements when hovered.
`--calcite-input-actions-icon-color-press`: Specifies the icon color of
Input's `clearable` and `number-button-type` elements when pressed.
 
`--calcite-input-loading-background-color`: Specifies the background
color of the `loading` element, when present.
`--calcite-input-loading-fill-color`: Specifies the fill color of the
`loading` element, when present.
 
`--calcite-input-prefix-background-color`: Specifies the background
color of the component's prefix, when present.
`--calcite-input-prefix-text-color`: Specifies the text color of the
prefix of the component's prefix, when present.

`--calcite-input-suffix-background-color`: Specifies the background
color of the component's suffix, when present.
`--calcite-input-suffix-text-color`: Specifies the text color of the
component's suffix, when present.

`--calcite-input-prefix-size`: Specifies the component's prefix width,
when present. _[existing]_
`--calcite-input-suffix-size`: Specifies the component's suffix width,
when present. _[existing]_