feat: add icon font files (#10996)
**Related Issue:**
[#56](https://github.com/Esri/calcite-ui-icons/issues/56)

## Summary
There has been multiple requests for a font using the Calcite UI icons.
This is filling that ask.
There are 3 font files and corresponding json files to organize and
standardize their codepoints.
Also contained in this folder is a codepoints.json for use when updating
the font (which will happen every 2 months in line with the release
schedule) to keep the codepoints increasing numerically and avoid
scrambling codepoints when adding new icons.

Thanks to contributions this process is now fully automated and will run automaticlly after every release. I'll keep a close eye on these files specifically after every release. 
Thanks 

I am also hoping to get the codepoints converted to Unicode and included
in the searchable ref page to make the font characters searchable. I’m
working with @macandcheese on that
[[here](https://github.com/ArcGIS/icon-explorer/issues/6)]

---------

Co-authored-by: Ben Elan <no-reply@benelan.dev>