feat(panel): control panel header action tokens (#11300)
## Summary

Per discussion on Teams. Allows the action menu in panel header
shadowDom to be controlled separately from other tokens without a reset
per customer request.

### Adds Tokens

--calcite-panel-header-action-background-color-hover: Specifies the
background color of the component's `calcite-action` items in the panel
header when hovered.
--calcite-panel-header-action-background-color-press: Specifies the
background color of the component's `calcite-action` items in the panel
header when pressed.
--calcite-panel-header-action-background-color: Specifies the background
color of the component's `calcite-action` items in the panel header.
--calcite-panel-header-action-indicator-color: Specifies the color of
the component's `calcite-action` items' indicator in the panel header.
--calcite-panel-header-action-text-color-press: Specifies the text color
of the component's `calcite-action` items in the panel header when
pressed.
--calcite-panel-header-action-text-color: Specifies the text color of
the component's `calcite-action` items in the panel header.