feat(dialog): add 'custom-content' slot (#11072)
**Related Issue:** #10323

## Summary
Deprecates the `"content"` slot in favor of a new `"custom-content"`
slot that will clearly describes the behavior that a user can expect.
This change should solve the issue of confusion that we've heard about
from internal feedback.

The description of the new slot would be: 
> **custom-content** - A slot for displaying custom content. Will
prevent the rendering of any default Dialog UI, except for `box-shadow`
and `corner-radius`.

deprecate(dialog): deprecate the `content` slot