fix(shell-panel): Fix visual issues in `displayMode="float-all"` (#11410)
**Related Issue:** #9926 

## Summary
Adds some styling to address the visual issues in Shell Panel.

![Screenshot 2025-01-29 at 6 58
09 PM](https://github.com/user-attachments/assets/957f812d-201e-42e5-b723-30d1362bd606)
![Screenshot 2025-01-29 at 6 57
27 PM](https://github.com/user-attachments/assets/15db515a-8573-44ee-ab9d-486aa3755165)



Note - I went down the rabbit hole of trying to conditionally pass
corner-radius to slotted Actions, Panels, etc. - it would involve a lot
of internal tokens and conditional states to handle various arrangements
of components. This could be something to look at in the future but I'm
not sure we want to handle this here.

![Screenshot 2025-01-29 at 6 54
39 PM](https://github.com/user-attachments/assets/e73c5c6a-deb2-4ad8-8733-ad716fbc0593)