fix(table): Fix `bordered` styling in Firefox (#11466)
**Related Issue:** #11464

## Summary
Adds a set of styling rules to address a Firefox-specific styling bug
with a "not best practice" disclaimer.
Also updates the Table "local demo" to use Dialog instead of Modal.

Note: in FF - the Table ends up being slightly taller because of this
adjustment - however, prior to this change, there was a greater
difference in height ...


| Before    | After |
| -------- | ------- |
| ![Screenshot 2025-02-05 at 3 11
36 PM](https://github.com/user-attachments/assets/86677e1e-2415-4244-886e-3671cbd01ee1)
| ![Screenshot 2025-02-05 at 3 10
24 PM](https://github.com/user-attachments/assets/159c5310-a8bc-4f94-a388-8ab52accef25)
|