ci: fix contributors action creating multiple PRs (#11693)
**Related Issue:** #11690

## Summary

The contributors action appends a uid to the branch name for some
reason, which results in a new PR being created every time a commit
lands on `dev`.


![image](https://github.com/user-attachments/assets/4c9773ec-9147-4d16-940a-122603896916)

I forked the action and made a small change that lets us rely on a
different action for the committing and PR creation/update process. We
can switch back to the upstream action if my PR is accepted and
installed.

ref: https://github.com/akhilmhdh/contributors-readme-action/pull/84