feat(combobox, combobox-item-group): add component tokens (#11623)
**Related Issue:** #7180 

## Summary

Adds following tokens in `combobox` component :

 `--calcite-combobox-icon-color`: Specifies the component's icon color.
`--calcite-combobox-icon-color-hover`: Specifies the component's icon
color when hovered.
`--calcite-combobox-background-color`: Specifies the background color of
the component's listbox.
`--calcite-combobox-input-border-color`: Specifies the border color of
the component's input.
`--calcite-combobox-input-background-color`: Specifies the background
color of the component's input.
`--calcite-combobox-input-height`: Specifies the height of the
component's input.
`--calcite-combobox-input-text-color`: Specifies the text color of the
component's input.

Adds following tokens in `combobox-item-group` component:

`--calcite-combobox-item-group-text-color`: Specifies the text color of
the component.
`--calcite-combobox-item-group-border-color`: Specifies the border color
of the component.