feat(combobox-item): add component tokens (#11645)
**Related Issue:** #7180 

## Summary

Adds following tokens in `combobox-item` component:

`--calcite-combobox-text-color`: Specifies the text color of the
component.
`--calcite-combobox-text-color-hover`: Specifies the text color of the
component when hovered.
`--calcite-combobox-item-background-color-active`: Specifies the
background color of the component when active.
`--calcite-combobox-item-background-color-hover`: Specifies the
background color of the component when hovered.
`--calcite-combobox-selected-icon-color`: Specifies the color of the
selected indicator icon.
`--calcite-combobox-description-text-color`: Specifies the text color of
the component's description & short-heading.
`--calcite-combobox-description-text-color-press`: Specifies the text
color of the component's description & short-heading when hovered.
`--calcite-combobox-heading-text-color`: Specifies the text color of the
component heading.