fix: outside click of a focus trap should not focus initial focused element (#11823)
**Related Issue:** #10682

## Summary

- We are running into an issue because we're using the `setReturnFocus`
to call `focusElement` on components.
- The focus trap logic calls `clickOutsideDeactivates` then
`setReturnFocus` which always gets passed the initial focused element.
- When `setReturnFocus` is not used and an outside click occurs, the
initial element is not focused.
- I'm not sure why this is different and would need some investigation.
- I think its because our setFocus methods requestAnimationFrame which
the default focus methods don't
- To work around this, I setup a flag to know when an outside click
occurs and to not do any calls to `focusElement`. This means when an
outside click occurs, no initial focus element is returned to focus by
the focus trap.

Discussion here:
https://github.com/focus-trap/focus-trap/discussions/1429#discussioncomment-12647625