fix(shell, sheet, dialog, modal): Fix for slotted component interaction (#12059)
**Related Issue:** #11657

## Summary
When components are slotted in Shell's named slots, ensure page content
is interactive outside the focus-trap.

In the issue, I mentioned using `allowsOutsideClick` - but decided on
`clickOutsideDeactivates` as the docs mention it takes precedence if
set.

I added tests on Shell as that is where the `embedded` prop is set on
the child components.

@driskull @Elijbet can you give this a look as I haven't worked on focus
trap that much.