feat(stepper): add component tokens (#12036)
**Related Issue:**
[#7180](https://github.com/Esri/calcite-design-system/issues/7180)

## Summary

Add `stepper` component tokens.

`--calcite-stepper-bar-gap`: Specifies the space between items in the
step bar.
`--calcite-stepper-bar-inactive-fill-color`: Specifies the fill color of
items in the step bar.
`--calcite-stepper-bar-active-fill-color`: Specifies the fill color of
active items in the step bar.
`--calcite-stepper-bar-complete-fill-color`: Specifies the fill color of
complete items in the step bar.
`--calcite-stepper-bar-error-fill-color`: Specifies the fill color of
error items in the step bar.