feat(input-number): add component tokens (#11961)
**Related Issue:** #7180 

## Summary

Adds the following tokens in input-number component

`--calcite-input-actions-background-color`: Specifies the background
color of the component's `clearable` and `number-button-type` elements.
`--calcite-input-actions-background-color-hover`: Specifies the
background color of the component's `clearable` and `number-button-type`
elements when hovered.
`--calcite-input-actions-background-color-press`: Specifies the
background color of the component's `clearable` and `number-button-type`
elements when pressed.
`--calcite-input-actions-icon-color`: Specifies the icon color of the
component's `clearable` and `number-button-type` elements.
`--calcite-input-actions-icon-color-hover`: Specifies the icon color of
the component's `clearable` and `number-button-type` elements when
hovered.
`--calcite-input-actions-icon-color-press`: Specifies the icon color of
the component's `clearable` and `number-button-type` elements when
pressed.
`--calcite-input-loading-background-color`: Specifies the background
color of the `loading` element.
`--calcite-input-loading-fill-color`: Specifies the fill color of the
`loading` element
`--calcite-input-number-background-color`: Specifies the background
color of the component.
`--calcite-input-number-border-color`: Specifies the border color of the
component.
`--calcite-input-number-corner-radius`: Specifies the border radius of
the component.
`--calcite-input-number-height`: Specifies the height of the component.
`--calcite-input-number-placeholder-text-color`: Specifies the text
color of the placeholder in the component.
`--calcite-input-number-text-color`: Specifies the text color of the
component.
`--calcite-input-number-text-color-focus`: Specifies the text color of
the component when focused.
`--calcite-input-prefix-background-color`: Specifies the background
color of the prefix element.
`--calcite-input-prefix-size`: Specifies the width of the prefix
element.
`--calcite-input-prefix-text-color`: Specifies the text color of the
prefix element.
`--calcite-input-suffix-background-color`: Specifies the background
color of the suffix element.
`--calcite-input-suffix-size`: Specifies the width of the suffix
element.
`--calcite-input-suffix-text-color`: Specifies the text color of the
suffix element.