feat(combobox): add `selectAll` toggle property (#11721)
**Related Issue:** #2311

## Summary
Provides a `Select All` checkbox in the `calcite-combobox` by default
for `selection-mode="multiple"` as a convenient shortcut for users to
quickly select or deselect all items in the list. The `indeterminate`
state for this checkbox is a visual and programmatic state that
indicates a checkbox is neither fully checked nor unchecked. It
represents a "partial selection" in scenario where a checkbox is
associated with a group of child checkboxes, and some (but not all) of
the child checkboxes are selected.

---------

Co-authored-by: Matt Driscoll <mdriscoll@esri.com>