chore: fail tests on unintentional warnings (#12162)
**Related Issue:** #10376

## Summary

Enforces test failures on console warnings.

Also adds a new test util (`logging.ts`) for mocking console methods,
useful for silencing or asserting log output.

### Notes

* The following components need to be refactored their associated
warnings:

    **`calcite-action-menu`**
    ```
@esri/calcite-components:test: [test:stable] Element
calcite-action-menu--lazy scheduled an update (generally because a
property was set) after an update completed, causing a new update to be
scheduled. This is inefficient and should be avoided unless the next
update can only be scheduled as a side effect of the previous update.
See https://lit.dev/msg/change-in-update for more information.
    ```
    **`calcite-tile-select`**
    ```
@esri/calcite-components:test: [test:stable] Element
calcite-tile-select--lazy scheduled an update (generally because a
property was set) after an update completed, causing a new update to be
scheduled. This is inefficient and should be avoided unless the next
update can only be scheduled as a side effect of the previous update.
See https://lit.dev/msg/change-in-update for more information.
    ```
    **`calcite-filter`**
    ```
@esri/calcite-components:test: [test:stable] No data was passed to the
filter function.
@esri/calcite-components:test: [test:stable] The data argument should be
an array of objects
    ```
* some `input`/`input-number` and date utils tests mock the console as
changes to value or type (from test helpers) cause the native
elements/APIs (i.e., Intl + dummy locales) to warn:
    ```
@esri/calcite-components:test: [test:stable] The specified value "" does
not conform to the required format. The format is "#rrggbb" where rr,
gg, bb are two-digit hexadecimal numbers.
    ```
* new failures helped spot some test issues:
  * carousel's `t9n` test using `calcite-combobox`
* add color token name fallback and font-weight support to `themed` and
updated the helper to throw an error instead of warning and using an
unexpected color token as a default
* improves custom theming story slightly (see
https://www.chromatic.com/test?appId=6266d45509d7eb004aa254fb&id=68360398aa24035be46b0e4f)
* tidies up `tile` themed tests that were covering core tokens, which
the test util does not support and was causing the 'unknown token'
warning (from above bullet) to be displayed
* removes `setupTests.ts` since global test setup is no longer needed