chore: tidy up theme test utils (#12349)
**Related Issue:** N/A

## Summary

This addresses some runtime errors showing up in demo pages. 

**Note**: custom component prop overrides will no longer work via the
`demo-theme` helper. This was technically not supported via the
`demo-theme` helper, but was working by being applied as a single string
to the element's `style` attribute. If this is needed, set token
customization can be set via the `style` attribute directly.