feat(input-text): add component tokens (#12242)
**Related Issue:** #7180 

## Summary

Adds following tokens in `input-text` component

`--calcite-input-action-background-color`: Specifies the background
color of the component's `clearable` element.
`--calcite-input-action-background-color-hover`: Specifies the
background color of the component's `clearable` element when hovered.
`--calcite-input-action-background-color-press`: Specifies the
background color of the component's `clearable` element when pressed.
`--calcite-input-action-icon-color`: Specifies the icon color of the
component's `clearable` & icon elements.
`--calcite-input-action-icon-color-hover`: Specifies the icon color of
the component's `clearable` & icon elements when hovered.
`--calcite-input-action-icon-color-press`: Specifies the icon color of
the component's `clearable` & icon elements when pressed.
`--calcite-input-loading-background-color`: Specifies the background
color of the `loading` element.
`--calcite-input-loading-fill-color`: Specifies the fill color of the
`loading` element.
`--calcite-input-prefix-background-color`: When `prefixText` is
provided, specifies the background color of the component's prefix
element.
`--calcite-input-prefix-size-x`: When `prefixText` is provided,
specifies the width of the component's prefix element.
`--calcite-input-prefix-text-color`: When `prefixText` is provided,
specifies the text color of the component's prefix element.
`--calcite-input-suffix-background-color`: When `suffixText` is
provided, specifies the background color of the component's suffix
element.
`--calcite-input-suffix-size-x`: When `suffixText` is provided,
specifies the width of the component's suffix element.
`--calcite-input-suffix-text-color`: When `suffixText` is provided,
specifies the color of the component's suffix element.
`--calcite-input-text-background-color`: Specifies the component's
background color.
`--calcite-input-text-border-color`: Specifies the component's border
color.
`--calcite-input-text-corner-radius`: Specifies the component's border
radius.
`--calcite-input-text-placeholder-text-color`: Specifies the component's
`placeholder` text color.
`--calcite-input-text-text-color`: Specifies the component's text color.
`--calcite-input-text-text-color-focus`: Specifies the component's text
color when focused.