feat(table, table-row, table-header, table-cell): Add component tokens (#11770)
**Related Issue:** #7180 

## Summary
Supports custom theming story for Table and related components:

![Screenshot 2025-06-02 at 3 46
25 PM](https://github.com/user-attachments/assets/a32e726d-5c6d-40dd-b3a8-eea848bcc32b)

Adds the following tokens:
### Table
`--calcite-table-border-color`: The border color of the component.
`--calcite-table-corner-radius`: The border color of the component.
`--calcite-table-shadow`: The shadow of the component.
`--calcite-table-number-cell-background-color`: The background color of
the component's `numbered` cell.
`--calcite-table-number-cell-text-color`: The text color of the
component's `numbered` cell.
`--calcite-table-selection-cell-background-color`: The background color
of the component's `selectionMode` cell.
`--calcite-table-selection-cell-background-color-selected`: The
background color of the component's `selectionMode` cell.
`--calcite-table-selection-cell-icon-color`: The icon color of the
component's `selectionMode` selection icon.
`--calcite-table-selection-cell-icon-color-selected`: The icon color of
the component's `selectionMode` selection icon when selected.
`--calcite-table-selection-chip-background-color`: The background color
of the component's `selectionMode` selection chip.
`--calcite-table-selection-chip-border-color`: The border color of the
component's `selectionMode` selection chip.
`--calcite-table-selection-chip-corner-radius`: The corner radius of the
component's `selectionMode` selection chip.
`--calcite-table-selection-chip-shadow`: The shadow of the component's
`selectionMode` selection chip.
`--calcite-table-selection-chip-text-color`: The text color of the
component's `selectionMode` selection chip.
`--calcite-table-selection-chip-selected-background-color`: The
background color of the component's `selectionMode` selection chip.
`--calcite-table-selection-chip-selected-border-color`: The border color
of the component's `selectionMode` selection chip.
`--calcite-table-selection-chip-selected-text-color`: The text color of
the component's `selectionMode` selection chip.

`--calcite-table-selection-out-of-view-chip-background-color`: The
background color of the component's `selectionMode` selection out of
view chip.
`--calcite-table-selection-out-of-view-chip-border-color`: The border
color of the component's `selectionMode` selection out of view chip.
`--calcite-table-selection-out-of-view-chip-corner-radius`: The corner
radius of the component's `selectionMode` selection out of view chip.
`--calcite-table-selection-out-of-view-chip-icon-color`: The icon color
of the component's `selectionMode` selection out of view chip.
`--calcite-table-selection-out-of-view-chip-shadow`: The shadow of the
component's `selectionMode` selection out of view chip.
`--calcite-table-selection-out-of-view-chip-text-color`: The text color
of the component's `selectionMode` selection out of view chip.
`--calcite-table-selection-dismiss-button-background-color-active`: The
background color of the component's `selectionMode` selection dismiss
chip when active.
`--calcite-table-selection-dismiss-button-background-color-hover`: The
background color of the component's `selectionMode` selection dismiss
chip when hovered.
`--calcite-table-selection-dismiss-button-background-color`: The
background color of the component's `selectionMode` selection dismiss
chip.
`--calcite-table-selection-dismiss-button-border-color-active`: The
border color of the component's `selectionMode` selection dismiss chip
when active.
`--calcite-table-selection-dismiss-button-border-color-hover`: The
border color of the component's `selectionMode` selection dismiss chip
when hovered.
`--calcite-table-selection-dismiss-button-border-color`: The border
color of the component's `selectionMode` selection dismiss chip.
`--calcite-table-selection-dismiss-button-corner-radius`: The corner
radius of the component's `selectionMode` selection dismiss chip.
`--calcite-table-selection-dismiss-button-shadow`: The shadow of the
component's `selectionMode` selection dismiss chip.
`--calcite-table-selection-dismiss-button-text-color-active`: The shadow
of the component's `selectionMode` selection dismiss chip when active.
`--calcite-table-selection-dismiss-button-text-color-hover`: The shadow
of the component's `selectionMode` selection dismiss chip when hovered.
`--calcite-table-selection-dismiss-button-text-color`: The shadow of the
component's `selectionMode` selection dismiss chip.
`--calcite-table-pagination-color`: Specifies the component's `pageSize`
pagination item color.
`--calcite-table-pagination-color-hover`: Specifies the component's
`pageSize` pagination item color when hovered or selected.
`--calcite-table-pagination-color-border-hover`: Specifies the
component's `pageSize` pagination item bottom border color when hovered.
`--calcite-table-pagination-color-border-active`: Specifies the
component's `pageSize` pagination item bottom border color when
selected.
`--calcite-table-pagination-background-color`: Specifies the component's
`pageSize` pagination item background color when active.
`--calcite-table-pagination-icon-color-background-hover`: Specifies the
component's `pageSize` pagination chevron item background color when
hovered.

### Table Header

`--calcite-table-header-background`: Specifies the background color of
the component.
`--calcite-table-header-background`: [Deprecated] Use the
`--calcite-table-header-background-color` property instead. Specifies
the background color of the component.
`--calcite-table-header-background-color`: Specifies the background
color of the component.
`--calcite-table-header-border-color`: Specifies the border color of the
component.
`--calcite-table-header-heading-color`: Specifies the heading color of
the component.
`--calcite-table-header-description-color`: Specifies the description
color of the component.

### Table Row
`--calcite-table-row-background`: [Deprecated] Use
`--calcite-table-row-background-color` instead. Specifies the background
color of the component.
`--calcite-table-row-background-color`: Specifies the background color
of the component.
`--calcite-table-row-background-color-striped`: The background color of
the component's `striped` rows, when specified.
`--calcite-table-row-background-color-selected`: The background color of
the component's `selected` rows, when specified.
`--calcite-table-row-selected-accent-color`: Specifies the selected
accent color of the component.
`--calcite-table-row-border-color`: Specifies the border color of the
component.

### Table Cell

`--calcite-table-cell-background`: [Deprecated] Use
`--calcite-table-cell-background-color` instead. Specifies the
background color of the component.
`--calcite-table-cell-background-color`: Specifies the background color
of the component.
`--calcite-table-cell-border-color`: Specifies the border color of the
component.
`--calcite-table-cell-text-color`: Specifies the text color of the
component.